Update AddIssuesField unit test
